#355 FloatingTexts bind to Player's position

+++ mdc/DataCenter/MDC/Data/Dtos/FloatingTextDto.cs
@@ -8,11 +8,11 @@ namespace MDC.Data.Dtos
 
         public string Level { get; set; }
 
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
-        public int Z { get; set; }
+        public double Z { get; set; }
 
         public DateTime CreatedDate { get; set; }
     }
