+++ server/plugins/MineParkCore/src/minepark/models/dtos/FloatingTextDto.php
@@ -7,11 +7,11 @@ class FloatingTextDto extends BaseDto
 
     public string $level;
 
-    public int $x;
+    public float $x;
 
-    public int $y;
+    public float $y;
 
-    public int $z;
+    public float $z;
 
     public string $createdDate;
 }
