+++ server/plugins/MineParkCore/src/minepark/models/dtos/PositionDto.php
@@ -5,10 +5,10 @@ class PositionDto extends BaseDto
 {
     public string $level;
 
-    public int $x;
+    public float $x;
 
-    public int $y;
+    public float $y;
 
-    public int $z;
+    public float $z;
 }
 ?>
